prover output

+++ examples/xdp/syn_prover.h
@@ -19,14 +19,14 @@ const unsigned long ETH_P_IPV6 = 34525;
 const unsigned long POW_THRESHOLD  = 4160749568; // k=32
 // const unsigned long POW_THRESHOLD  = 4227858432; // k=64
 
-const int POW_ITERS = 100;
+const unsigned short POW_ITERS = 100;
 
 struct event {
 	unsigned long long start_ts;
 	unsigned long long end_ts;
 	unsigned long best_nonce;
 	unsigned long best_hash;
-	unsigned int hash_iters;
+	unsigned short hash_iters;
 };
 
 #endif /* __SYN_PROVER_H */
